+++ static/js/redux/actions/search_actions.jsx
@@ -17,7 +17,6 @@ import { getActiveTimetableCourses, getCurrentSemester } from "../state";
 import { getCourseSearchEndpoint } from "../constants/endpoints";
 import { getUserSavedTimetables, saveTimetable } from "./user_actions";
 import { nullifyTimetable } from "./timetable_actions";
-import * as ActionTypes from "../constants/actionTypes";
 import { fetchCourseClassmates } from "./modal_actions";
 import { getSemester } from "./school_actions";
 import { alertsActions, explorationModalActions } from "../state/slices";
