+++ static/js/redux/state/index.tsx
@@ -41,7 +41,7 @@ import termsOfServiceModal from "./slices/termsOfServiceModalSlice";
 import termsOfServiceBanner from "./slices/termOfServiceBannerSlice";
 import userAcquisitionModal from "./slices/userAcquisitionModalSlice";
 import registrar from "./slices/registrarSlice";
-import entities, * as fromEntities from "./entities_reducer";
+import entities, * as fromEntities from "./slices/entitiesSlice";
 import { Slot, Timetable } from "../constants/commonTypes";
 
 export const reducers = {
@@ -108,6 +108,7 @@ export const getCoursesFromSlots = (state: RootState, slots: Slot[]) =>
   fromEntities.getCoursesFromSlots(state.entities, slots);
 
 export const getMaxTTEndHour = createSelector(
+  // @ts-ignore
   [getActiveDenormTimetable],
   fromEntities.getMaxEndHour
 );
