+++ static/js/redux/ui/containers/modals/course_modal_body_container.jsx
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import CourseModalBody from "../../modals/course_modal_body";
 import { getActiveTimetable, getCurrentSemester } from "../../../state";
-import { getSectionTypeToSections } from "../../../state/entities_reducer";
+import { getSectionTypeToSections } from "../../../state/slices/entitiesSlice";
 import { fetchCourseInfo, react } from "../../../actions/modal_actions";
 import { saveSettings } from "../../../actions/user_actions";
 import { getSchoolSpecificInfo } from "../../../constants/schools";
